Adding new Intercepting layer for consent management
